WIP: Fix circular import issue between validation.py and bitdict.py
- Created new factory.py module to break circular dependency
- Moved BitDict implementation and ABC from bitdict.py to factory.py
- Updated validation.py to use dependency injection instead of direct import
- Modified all import statements across the codebase
- Fixed nested bitdict validation to properly handle default values and subtypes
- Updated tests to expect correct behavior with defaults added
- All 147 tests now pass
- Circular import completely resolved
