+++ bitdict/__init__.py
@@ -1,6 +1,6 @@
 """bitdict package."""
 
-from bitdict.bitdict import BitDictABC, bitdict_factory
+from bitdict.factory import BitDictABC, bitdict_factory
 from bitdict.markdown import generate_markdown_tables
 
 __all__ = ["bitdict_factory", "generate_markdown_tables", "BitDictABC"]
