+++ bitdict/bitdict.py
@@ -28,919 +28,7 @@ Usage:
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from copy import deepcopy
-from types import MappingProxyType
-from typing import Any, Callable, Generator
+# Import the factory and ABC from the new factory module
+from .factory import BitDictABC, bitdict_factory
 
-from .validation import check_overlapping, validate_property_config
-
-
-def _calculate_total_width(cfg: dict[str, Any]) -> int:
-    """
-    Calculates the total bit width based on the configuration.
-
-    This is a helper function that calculates the total number of bits
-    required to store all the bit fields defined in the configuration.
-
-    Args:
-        cfg: The configuration dictionary.
-
-    Returns:
-        int: The total bit width.
-    """
-    max_bit = 0
-    for prop_config in cfg.values():
-        end_bit = prop_config["start"] + prop_config["width"]
-        max_bit = max(max_bit, end_bit)
-    return max_bit
-
-
-def _is_valid_value(value: int | bool, prop_config: dict[str, Any]) -> bool:
-    """Checks if a value is valid according to the 'valid'
-    key in the property configuration."""
-    # pylint: disable=duplicate-code
-    if "valid" not in prop_config:
-        return True
-
-    valid_config = prop_config["valid"]
-    if "value" in valid_config and value in valid_config["value"]:
-        return True
-
-    if "range" in valid_config:
-        for r in valid_config["range"]:
-            if value in range(*r):
-                return True
-
-    return False
-
-
-class BitDictABC(ABC):
-    """Abstract base class for BitDict"""
-
-    _config: MappingProxyType[str, Any]
-    _total_width: int
-    subtypes: dict[str, list[type[BitDictABC] | None]]
-    _total_width: int
-    title: str
-
-    @abstractmethod
-    def __init__(
-        self,
-        value: int | bytes | bytearray | dict[str, Any] | None = None,
-        ignore_unknown: bool = True,
-    ) -> None:
-        """Initializes a BitDict instance with a specified value."""
-
-    @abstractmethod
-    def __getitem__(self, key: str) -> bool | int | BitDictABC:
-        """Retrieves the value associated with the given key."""
-
-    @abstractmethod
-    def __setitem__(self, key: str, value: bool | int) -> None:
-        """Sets the value of a property within the BitDict."""
-
-    @abstractmethod
-    def __len__(self) -> int:
-        """Returns the total width of the bit dictionary."""
-
-    @abstractmethod
-    def __contains__(self, key: str) -> bool:
-        """Check if a property exists within this BitDict or its nested BitDict"""
-
-    @abstractmethod
-    def __iter__(
-        self,
-    ) -> Generator[tuple[str, bool | int | BitDictABC], Any, None]:
-        """BitDict iterator to traverse properties in LSB to MSB order"""
-
-    @abstractmethod
-    def __repr__(self) -> str:
-        """Returns a string representation of the BitDict"""
-
-    @abstractmethod
-    def __str__(self) -> str:
-        """Returns a string representation of the BitDict"""
-
-    @abstractmethod
-    def _get_value(self) -> int:
-        """Returns the value of the BitDict"""
-
-    @abstractmethod
-    def _set_parent(self, parent: BitDictABC, key: str) -> None:
-        """Sets the parent BitDict and key for this BitDict"""
-
-    @abstractmethod
-    def _set_value(self, value: int) -> None:
-        """Sets the value of the BitDict"""
-
-    @classmethod
-    @abstractmethod
-    def assign_verification_function(
-        cls, verification_function: Callable[[BitDictABC], bool]
-    ) -> None:
-        """Assigns a verification function to the BitDict"""
-
-    @abstractmethod
-    def clear(self) -> None:
-        """Clears all bit fields to their default values."""
-
-    @classmethod
-    @abstractmethod
-    def get_config(cls) -> MappingProxyType[str, Any]:
-        """Returns the configuration dictionary for the BitDict."""
-
-    @abstractmethod
-    def inspect(self) -> dict[str, Any]:
-        """Inspects the current state of the BitDict and returns a dictionary."""
-
-    @abstractmethod
-    def reset(self) -> None:
-        """Resets all bit fields to their default values."""
-
-    @abstractmethod
-    def set(self, value: int | dict[str, Any], ignore_unknown: bool = True) -> None:
-        """Sets the BitDict to the provided integer value
-        or a dictionary for multi-field updates."""
-
-    @abstractmethod
-    def update(self, data: dict[str, Any]) -> None:
-        """Updates the BitDict with the provided data."""
-
-    @abstractmethod
-    def to_bytes(self) -> bytes:
-        """Converts the BitDict to a byte representation."""
-
-    @abstractmethod
-    def to_int(self) -> int:
-        """Converts the BitDict to an integer representation."""
-
-    @abstractmethod
-    def to_json(self) -> dict[str, Any]:
-        """Converts the BitDict to a JSON-compatible dictionary."""
-
-    @abstractmethod
-    def valid(self) -> bool:
-        """Checks if the bit dictionary is valid."""
-
-    @abstractmethod
-    def verify(self) -> bool:
-        """Verifies the integrity of the bit dictionary."""
-
-
-def bitdict_factory(  # pylint: disable=too-many-statements
-    config: dict[str, Any], name: str = "BitDict", title: str = "BitDict"
-) -> type[BitDictABC]:
-    """
-    Factory function to create BitDict classes based on a configuration.
-
-    This factory function dynamically creates a new class (subclass of BitDict)
-    based on the provided configuration dictionary. The generated class allows
-    you to define and access bit fields within an integer value, similar to
-    a struct in C.
-
-    Args:
-        config (dict): A dictionary defining the bit field structure.
-            The dictionary keys are the names of the bit fields, and the values
-            are dictionaries with the following keys:
-
-            * `start` (int): The starting bit position (LSB = 0).
-            * `width` (int): The width of the bit field in bits.
-            * `type` (str): The data type of the bit field.
-              Can be one of: 'bool', 'uint', 'int', 'reserved', 'bitdict'.
-            * `default` (optional): The default value for the bit field.
-              If not provided, defaults to False for 'bool', 0 for 'uint' and 'int'.
-            * `subtype` (list of dict, optional):  Required for 'bitdict' type.
-              Ignored for other types.
-              A list of sub-bitdict configurations to select from based
-              on the value of the selector property.
-            * `selector` (str, optional): Required for 'bitdict' type.
-              Ignored for other types
-              The name of the property used to select the active sub-bitdict.
-              This property must be of type 'bool' or 'uint' and have a width <= 16.
-
-        name (str, optional): The name of the generated class.
-            Defaults to "BitDict". Must be a valid python identifier.
-        title (str, optional): The title of the generated class. Used in generating
-            the markdown documentation with `config_to_markdown`. Defaults to "BitDict".
-
-    Returns:
-        type: A new class that represents the bit field structure.
-
-    Raises:
-        ValueError: If the configuration is invalid (e.g., overlapping
-            bit fields, invalid property names, missing required keys).
-        TypeError: If the config is not a dictionary.
-
-    Example:
-        ```python
-        config = {
-            'enabled': {'start': 0, 'width': 1, 'type': 'bool'},
-            'mode': {'start': 1, 'width': 2, 'type': 'uint'},
-            'value': {'start': 3, 'width': 5, 'type': 'int'},
-        }
-        MyBitDict = bitdict_factory(config, "MyBitDict")
-
-        # Create an instance of the generated class
-        bd = MyBitDict()
-
-        # Set and access bit fields
-        bd['enabled'] = True
-        bd['mode'] = 2
-        bd['value'] = -5
-
-        # Get the integer representation
-        value = bd.to_int()
-
-        # Create an instance from an integer
-        bd2 = MyBitDict(value)
-        ```
-    """
-    if not isinstance(config, dict):  # type: ignore[runtime safety]
-        raise TypeError("config must be a dictionary")
-    if not name.isidentifier():
-        raise ValueError("Invalid class name")
-
-    # Subtype classes are stored in a dictionary for recursive creation.
-    subtype_lists: dict[str, list[type[BitDictABC] | None]] = {}
-    _title: str = title
-
-    validate_property_config(config, subtype_lists)  # Initial validation of top level.
-    check_overlapping(config)
-    total_width = _calculate_total_width(config)
-
-    class BitDict(BitDictABC):
-        """
-        A dynamic bit-field dictionary for structured data manipulation.
-
-        BitDict provides a flexible way to define and interact with data structures
-        where individual fields occupy specific bit ranges. It supports a variety of
-        data types including integers (signed and unsigned), booleans, and nested
-        BitDict instances, enabling complex data layouts to be easily managed.
-
-        Key Features:
-        - **Dynamic Configuration:**  BitDict classes are dynamically created based on a
-          provided configuration, defining the structure and properties of the bit field.
-        - **Sub-BitDict Support:** Allows nesting of BitDict instances, enabling hierarchical
-          data structures with conditional sub-fields based on selector values.
-        - **Type Handling:** Enforces type and range checking for property assignments,
-          ensuring data integrity.
-        - **Data Conversion:** Supports conversion to and from integers, bytes, and
-          JSON-compatible dictionaries.
-        - **Iteration:** Provides an iterator to traverse properties in LSB to MSB order.
-
-        Use Cases:
-        - Parsing and generating binary data formats (e.g., network packets, file formats).
-        - Representing hardware registers and memory-mapped I/O.
-        - Implementing data structures with specific bit-level packing requirements.
-
-        Example:
-        ```python
-        config = {
-            "enabled": {"type": "bool", "start": 0, "width": 1, "default": False},
-            "mode": {"type": "uint", "start": 1, "width": 2, "default": 0},
-            "value": {"type": "int", "start": 3, "width": 5, "default": 0},
-        }
-        MyBitDict = bitdict_factory("MyBitDict", config)
-        my_dict = MyBitDict()
-        my_dict["enabled"] = True
-        my_dict["mode"] = 2
-        my_dict["value"] = -5
-        print(my_dict.to_json())
-        ```
-        """
-
-        _config: MappingProxyType[str, Any] = MappingProxyType(deepcopy(config))
-        subtypes: dict[str, list[type[BitDictABC] | None]] = subtype_lists
-        _total_width: int = total_width
-        title: str = _title
-        __name__: str = name
-        verification_function: Callable[[BitDictABC], bool] = staticmethod(  # type: ignore[misc]
-            lambda _: True  # type: ignore[misc]
-        )
-        __slots__: tuple[str, ...] = (
-            "_value",
-            "_subbitdicts",
-            "_parent",
-            "_parent_key",
-        )
-
-        def __init__(
-            self,
-            value: int | bytes | bytearray | dict[str, Any] | None = None,
-            ignore_unknown: bool = True,
-        ) -> None:
-            """Initializes a BitDict instance with a specified value.
-            The initial value can be provided as an integer,
-            bytes/bytearray, or a dictionary. If no value is provided,
-            the BitDict is initialized to its default state (all bits
-            zeroed).
-
-                value (int | bytes | bytearray | dict[str, Any] | None,
-                optional): The initial value for the BitDict.
-                    - If `int`, the BitDict is set to this integer value.
-                      Value must be within the representable range.
-                    - If `bytes` or `bytearray`, the BitDict is initialized
-                      from the big-endian representation of these bytes. The
-                      length of the byte sequence must be appropriate for the
-                      BitDict's total width.
-                    - If `dict`, the BitDict is initialized using the
-                      dictionary's key-value pairs.
-                    - If `None`, the BitDict is initialized to its default
-                      state. Defaults to None.
-
-                ignore_unknown (bool, optional): If set to True, unknown properties
-                    in the initialization dictionary will be ignored. If False
-                    an unknown property will raise a KeyError. Defaults to True.
-
-                TypeError: If `value` is not one of the supported types
-                    (int, bytes, bytearray, dict, None).
-                ValueError:
-                    - If `value` is an integer that exceeds the maximum or
-                      falls below the minimum representable value given the
-                      BitDict's total width.
-                    - If `value` is a bytes or bytearray object whose length
-                      is incompatible with the BitDict's total width.
-            """
-            self._value = 0
-            # Instances of subbitdicts
-            self._subbitdicts: dict[str, list[BitDictABC | None]] = {}
-
-            # Identification of this BitDict in a parent BitDict
-            # These are set by _parent_config() when this BitDict is a sub-bitdict.
-            # They are used in _update_parent() to update the parent BitDict
-            # when this BitDict changes.
-            self._parent: BitDictABC | None = None
-            self._parent_key: str | None = None
-
-            # Set to defaults
-            if value is None:
-                self.reset()
-            elif isinstance(value, int):
-                if value >= (1 << self._total_width):
-                    raise ValueError(
-                        f"Integer value {value} exceeds maximum"
-                        f" value for bit width {self._total_width}"
-                    )
-                if value < -(1 << (self._total_width - 1)):
-                    raise ValueError(
-                        f"Integer value {value} exceeds minimum"
-                        f"value for bit width {self._total_width}"
-                    )
-                self.set(value)
-            elif isinstance(value, (bytes, bytearray)):
-                if len(value) > (self._total_width + 7) // 8:  # +7 to round up
-                    raise ValueError(
-                        f"Bytes object too long for bit width {self._total_width}"
-                    )
-                # Convert bytes to integer (big-endian)
-                self.set(int.from_bytes(value, "big"))
-            elif isinstance(value, dict):  # type: ignore[misc]
-                self.set(
-                    value, ignore_unknown
-                )  # Use update to handle defaults and type checking
-            else:
-                raise TypeError(
-                    "Invalid initialization type: must be None, int, bytes, bytearray, or dict"
-                )
-
-        def __getitem__(self, key: str) -> bool | int | BitDictABC:
-            """
-            Retrieves the value associated with the given key.
-            The key corresponds to a property defined in the BitDict's configuration.
-            The type of the returned value depends on the property's type:
-            - 'bool': Returns a boolean value.
-            - 'int': Returns a signed integer value (two's complement).
-            - 'uint': Returns an unsigned integer value.
-            - 'bitdict': Returns a sub-BitDict, selected by the value of another property.
-            Args:
-                key: The name of the property to retrieve.
-            Returns:
-                The value of the property, with the type depending on the property's configuration.
-                Can be a bool, int, or BitDict.
-            Raises:
-                KeyError: If the key is not a valid property in the configuration.
-                AssertionError: If the selector value for a 'bitdict' type is not an integer,
-                        or if an unknown property type is encountered.
-            """
-            if key not in self._config:
-                raise KeyError(f"Invalid property: {key}")
-
-            prop_config = self._config[key]
-            start = prop_config["start"]
-            width = prop_config["width"]
-            mask = (1 << width) - 1
-            raw_value = (self._value >> start) & mask
-
-            if prop_config["type"] == "bool":
-                return bool(raw_value)
-
-            if prop_config["type"] == "uint":
-                return raw_value
-
-            if prop_config["type"] == "int":
-                # Two's complement conversion if the highest bit is set
-                return (
-                    raw_value - (1 << width)
-                    if raw_value & (1 << (width - 1))
-                    else raw_value
-                )
-
-            if prop_config["type"] == "bitdict":
-                selector_value: bool | int | BitDictABC = self[prop_config["selector"]]
-                assert isinstance(selector_value, int), "Selector must be an integer"
-                bd: BitDictABC = self._get_subbitdict(key, selector_value)
-                return bd
-
-            assert False, f"Unknown property type: {prop_config['type']}"
-
-        def __setitem__(self, key: str, value: bool | int) -> None:
-            """Sets the value of a property within the BitDict.
-            Args:
-                key (str): The name of the property to set.
-                value (bool | int): The value to set the property to.  Must be
-                a boolean or integer, depending on the property's type.
-            Raises:
-                KeyError: If the given key is not a valid property in the BitDict's configuration.
-                ValueError: If attempting to set a reserved property, or if the provided value
-                is out of the allowed range for the property.
-                TypeError: If the provided value is not of the expected type (boolean or integer)
-                for the property.
-            """
-            if key not in self._config:
-                raise KeyError(f"Invalid property: {key}")
-
-            prop_config = self._config[key]
-            start = prop_config["start"]
-            width = prop_config["width"]
-            mask = (1 << width) - 1
-
-            match prop_config["type"]:
-                case "bool":
-                    if not isinstance(value, (bool, int)):  # type: ignore[misc]
-                        raise TypeError(
-                            f"Expected boolean or integer value for property '{key}'"
-                        )
-                    value = 1 if value else 0
-                case "uint":
-                    if not isinstance(value, int):  # type: ignore[misc]
-                        raise TypeError(f"Expected integer value for property '{key}'")
-                    if not 0 <= value < (1 << width):
-                        raise ValueError(
-                            f"Value {value} out of range for property '{key}'"
-                        )
-                case "int":
-                    if not isinstance(value, int):  # type: ignore[misc]
-                        raise TypeError(f"Expected integer value for property '{key}'")
-                    if not -(1 << (width - 1)) <= value < (1 << (width - 1)):
-                        raise ValueError(
-                            f"Value {value} out of range for property '{key}'"
-                        )
-                    # Convert to two's complement representation
-                    if value < 0:
-                        value = (1 << width) + value
-                case "bitdict":
-                    # Set the sub-bitdict value.
-                    selector_value = self[prop_config["selector"]]
-                    assert isinstance(
-                        selector_value, int
-                    ), "Selector must be an integer"
-                    bd: BitDictABC = self._get_subbitdict(key, selector_value)
-                    bd.set(value)
-                    value = bd.to_int()
-                case _:
-                    assert False, f"Unknown property type: {prop_config['type']}"
-
-            # If the property is a selector then the sub-bitdict
-            # changes and we need to update the value
-            # Note that if the newly selected BitDict was previously
-            # defined then that value will be used
-            # or else it will be the default for the new BitDict.
-            # It will not maintain the same numeric value.
-            # This is important to call out in the user documentation.
-            if "_bitdict" in prop_config:
-                bd = self._get_subbitdict(prop_config["_bitdict"], value)
-                bdc = self._config[prop_config["_bitdict"]]
-                _mask = (1 << bdc["width"]) - 1
-                _start = bdc["start"]
-                self._value &= ~(_mask << _start)
-                self._value |= (bd.to_int() & _mask) << _start
-
-            # Clear the bits for this property, then set the new value
-            self._value &= ~(mask << start)
-            self._value |= (value & mask) << start
-
-            # If this BitDict is a sub-bitdict then update the parent
-            if self._parent is not None:
-                self._update_parent()
-
-        def __len__(self) -> int:
-            """
-            Returns the total width of the bit dictionary, representing the number
-            of bits it can store.
-            Returns:
-                int: The total width (number of bits) of the bit dictionary.
-            """
-
-            return self._total_width
-
-        def __contains__(self, key: str) -> bool:
-            """Check if a property exists within this BitDict or its nested BitDicts.
-            This method checks for the existence of a given key in the BitDict's configuration.
-            It considers the current selector state, meaning that if a property exists only within
-            a deselected subtype, this method will return False.
-            Args:
-                key: The property name (key) to check for.
-            Returns:
-                True if the property exists and is accessible given the current selector state,
-                False otherwise.
-            """
-            retval: bool = key in self._config
-            if not retval:
-                for k, bd in self._config.items():
-                    if bd["type"] == "bitdict":
-                        bdi: bool | int | BitDictABC = self[k]
-                        assert not isinstance(bdi, int), "Expecting BitDict type."
-                        retval = retval or key in bdi
-                        if retval:
-                            break
-            return retval
-
-        def __iter__(self) -> Generator[tuple[str, bool | BitDictABC | int], Any, None]:
-            """Iterates over the BitDict, yielding (name, value) pairs for each
-            non-reserved field.
-            Yields:
-                Generator[tuple[Any, bool | Any | BitDict | None], Any, None]:
-                A generator that yields tuples of (name, value), where name is the
-                field name and value is the corresponding value in the BitDict.
-                The values can be of type bool, Any, BitDict, or None.
-            """
-            sps = sorted(self._config.items(), key=lambda item: item[1]["start"])
-            for name, _ in sps:
-                yield name, self[name]
-
-        def __repr__(self) -> str:
-            """
-            Return a string representation of the BitDict object.
-            The string representation includes the class name and a JSON-like
-            representation of the BitDict's contents, obtained via the `to_json()` method.
-            Returns:
-                str: A string representation of the BitDict.
-            """
-
-            return f"{self.__class__.__name__}({self.to_json()})"
-
-        def __str__(self) -> str:
-            """
-            Returns a string representation of the BitDict object.
-
-            This method converts the BitDict object to its JSON representation
-            and then returns the string representation of that JSON object.
-
-            Returns:
-                str: A string representation of the BitDict object in JSON format.
-            """
-
-            return str(self.to_json())
-
-        def _get_subbitdict(self, key: str, selector_value: int) -> BitDictABC:
-            """Retrieves a sub-BitDict associated with a given key and selector value.
-            If the sub-BitDict does not already exist, it is created, initialized,
-            and stored for future access.
-            Args:
-                key: The key associated with the sub-BitDict.  This corresponds to a
-                 property defined in the BitDict's configuration.
-                selector_value: The selector value used to identify the specific
-                        sub-BitDict within the list of possible sub-BitDicts
-                        for the given key.  This value is typically derived
-                        from another property acting as a selector.
-            Returns:
-                The sub-BitDict associated with the given key and selector value.
-                The returned BitDict is guaranteed to exist.
-            Raises:
-                AssertionError: If the subtype class has not been created for the
-                        given selector value, or if the created sub-BitDict
-                        is None.
-            """
-
-            prop_config = self._config[key]
-            if key not in self._subbitdicts:
-                width = self._config[prop_config["selector"]]["width"]
-                self._subbitdicts[key] = [None] * 2**width
-            sdk: list[BitDictABC | None] = self._subbitdicts[key]
-            if sdk[selector_value] is None:
-                if selector_value >= len(self.subtypes[key]):
-                    raise IndexError(
-                        "Subtype class not created for selector"
-                        f" {prop_config['selector']} at index {selector_value}"
-                    )
-                bdtype: type[BitDictABC] | None = self.subtypes[key][selector_value]
-                assert bdtype is not None, "Subtype class not created!"
-                nbd = bdtype()
-                nbd._set_parent(self, key)  # pylint: disable=protected-access
-                sdk[selector_value] = nbd
-            retval: BitDictABC | None = sdk[selector_value]
-            assert retval is not None, "Subtype class not created!"
-            return retval
-
-        def _get_value(self) -> int:
-            """Returns the integer value representing the BitDict."""
-            return self._value
-
-        def _set_parent(self, parent: BitDictABC, key: str) -> None:
-            """Sets the parent BitDict and the key associated with this BitDict in the parent.
-            Args:
-                parent: The parent BitDict.
-                key: The key associated with this BitDict in the parent.
-            """
-
-            self._parent = parent
-            self._parent_key = key
-
-        def _set_value(self, value: int) -> None:
-            """Sets the integer value representing the BitDict.
-            This method is used internally to set the BitDict's value
-            without triggering updates to the parent BitDict.
-            Args:
-                value: The integer value to set.
-            """
-            self._value = value
-
-        def _update_parent(self) -> None:
-            """Updates the parent BitField with the current value of this BitField.
-            This method assumes that the parent BitField, as well as the key
-            associated with this BitField within the parent's configuration,
-            are properly set. It calculates a mask based on the width defined
-            in the parent's configuration for this BitField, clears the
-            corresponding bits in the parent's value, and then sets those bits
-            to the current value of this BitField.
-            """
-
-            assert self._parent is not None, "Parent not set"
-            assert self._parent_key is not None, "Parent key not set"
-            pc = self._parent._config[  # pylint: disable=protected-access
-                self._parent_key
-            ]
-            mask = (1 << pc["width"]) - 1
-            start = pc["start"]
-            _value = self._parent._get_value()  # pylint: disable=protected-access
-            _value &= ~(mask << start)
-            _value |= (self.to_int() & mask) << start
-            self._parent._set_value(_value)  # pylint: disable=protected-access
-
-        @classmethod
-        def assign_verification_function(
-            cls, verification_function: Callable[[BitDictABC], bool]
-        ) -> None:
-            """Assigns a verification function to the BitDict.
-            The verification function is called by the `verify()` method to check the
-            validity of the BitDict's properties.
-            Args:
-                verification_function: A function that takes a BitDict instance as input
-                    and returns a boolean indicating whether the BitDict's properties are valid.
-            """
-            cls.verification_function = verification_function
-
-        def clear(self) -> None:
-            """Clears the bit dictionary, setting all bits to 0."""
-
-            self.set(0)
-
-        @classmethod
-        def get_config(cls) -> MappingProxyType[str, Any]:
-            """Returns the configuration settings for the BitDict class.
-            The configuration is stored in a MappingProxyType, which provides
-            a read-only view of the underlying dictionary. This prevents
-            accidental modification of the configuration after the class
-            has been initialized.
-            Returns:
-                MappingProxyType[str, Any]: A read-only mapping containing the
-                configuration settings.
-            """
-
-            return cls._config
-
-        def inspect(self) -> dict[str, Any]:
-            """Inspects the BitDict and returns a dictionary of properties with invalid values."""
-            invalid_props: dict[str, Any] = {}
-            for prop_name, prop_config in self._config.items():
-                if prop_config["type"] == "bitdict":
-                    selector_value = self[prop_config["selector"]]
-                    assert isinstance(
-                        selector_value, int
-                    ), "Selector must be an integer"
-                    if not _is_valid_value(
-                        selector_value, self._config[prop_config["selector"]]
-                    ):
-                        invalid_props[prop_config["selector"]] = selector_value
-                    else:
-                        sub_bitdict = self._get_subbitdict(prop_name, selector_value)
-                        sub_invalid_props = sub_bitdict.inspect()
-                        if sub_invalid_props:
-                            invalid_props[prop_name] = sub_invalid_props
-                else:
-                    value = self[prop_name]
-                    assert isinstance(
-                        value, (int, bool)
-                    ), "Value must be an integer or boolean"
-                    if not _is_valid_value(value, prop_config):
-                        invalid_props[prop_name] = self[prop_name]
-            return invalid_props
-
-        def reset(self) -> None:
-            """Resets the BitDict to its default values.
-            Iterates through the properties defined in the configuration.
-            If a property is a reserved type, it is skipped.
-            If a property is a nested BitDict, its reset method is called recursively.
-            Otherwise, the property is set to its default value as specified in the configuration.
-            """
-
-            for prop_name, prop_config in self._config.items():
-                if prop_config["type"] == "bitdict":
-                    selector_value = self[prop_config["selector"]]
-                    assert isinstance(
-                        selector_value, int
-                    ), "Selector must be an integer"
-                    self._get_subbitdict(prop_name, selector_value).reset()
-                else:
-                    self[prop_name] = prop_config["default"]
-
-        def set(self, value: int | dict[str, Any], ignore_unknown: bool = True) -> None:
-            """Sets the value of the BitDict.
-            The value can be set in two ways:
-            1.  As an integer: In this case, the integer value is assigned to the
-                underlying integer representation of the BitDict.  A ValueError is
-                raised if the integer is outside the allowed range, given the
-                configured bit width. Note the integer must be >= 0.
-            2.  As a dictionary: In this case, the dictionary is treated as a set
-                of property values to be assigned to the BitDict.  The keys of the
-                dictionary correspond to the property names defined in the BitDict's
-                configuration.  If a property is present in the dictionary, its
-                value is assigned to the corresponding sub-BitDict or bit field.
-                If a property is not present in the dictionary but has a "default"
-                value specified in the configuration, the default value is assigned.
-            Args:
-                value: The value to set.  Can be an integer or a dictionary of
-                property values.
-                ignore_unknown: If True, unknown properties in the dictionary
-                will be ignored. If False, a KeyError will be raised for unknown
-                properties.
-            Raises:
-                ValueError: If the integer value is outside the allowed range for
-                the configured bit width.
-                AssertionError: If the selector value is not an integer when setting
-                a sub-BitDict.
-            """
-            # pylint: disable=too-many-branches
-
-            if isinstance(value, dict):
-                if ignore_unknown:
-                    # If ignoring unknown field names, set the fields that
-                    # are in the config and set the defaults for the rest.
-                    for prop_name, prop_config in self._config.items():
-                        if prop_name in value:
-                            self[prop_name] = value[prop_name]
-                        elif "default" in prop_config:
-                            self[prop_name] = prop_config["default"]
-                else:
-                    # If not ignoring unknown field names, set the fields
-                    # that are in the config and raise an error for the rest.
-                    # Set the fields not in the value to their default values.
-                    defaults = set(self._config.keys())
-                    for v_name in value:
-                        if v_name not in self._config:
-                            raise KeyError(f"Unknown property: {v_name}")
-                        self[v_name] = value[v_name]
-                        defaults.discard(v_name)
-                    for prop_name in defaults:
-                        prop_config = self._config[prop_name]
-                        if "default" in prop_config:
-                            self[prop_name] = prop_config["default"]
-            else:
-                if value >= (1 << self._total_width):
-                    raise ValueError(
-                        f"Integer value {value} exceeds maximum"
-                        f" value for bit width {self._total_width}"
-                    )
-                if value < 0:
-                    raise ValueError(f"Integer must be non-negative, got {value}")
-                self._value: int = value
-
-                # Must set sub-bitdicts after setting the main value.
-                for prop_name, prop_config in self._config.items():
-                    if prop_config["type"] == "bitdict":
-                        selector_value = self[prop_config["selector"]]
-                        assert isinstance(
-                            selector_value, int
-                        ), "Selector must be an integer"
-                        bd_value = (value >> prop_config["start"]) & (
-                            (1 << prop_config["width"]) - 1
-                        )
-                        self._get_subbitdict(prop_name, selector_value).set(bd_value)
-
-        def update(self, data: dict[str, Any]) -> None:
-            """Update the BitDict with values from another dictionary.
-            Args:
-                data (dict[str, Any]): A dictionary containing the keys and values to update
-                in the BitDict.  Keys must be strings and values must be convertible
-                to integers within the BitDict's bit_length.
-            Raises:
-                TypeError: If `data` is not a dictionary.
-                ValueError: If a value in `data` cannot be represented within the
-                BitDict's bit_length.
-            """
-
-            if not isinstance(data, dict):  # type: ignore[misc]
-                raise TypeError("update() requires a dictionary")
-            for key, value in data.items():
-                self[key] = value  # Use __setitem__ for type/range checking
-
-        def to_bytes(self) -> bytes:
-            """Convert the bit dictionary to a byte string.
-            The resulting byte string represents the underlying integer value
-            of the bit dictionary in big-endian byte order. The length of the
-            byte string is determined by the total width of the bit dictionary,
-            rounded up to the nearest whole byte.
-            Returns:
-                bytes: A byte string representing the bit dictionary's value.
-            """
-
-            num_bytes = (self._total_width + 7) // 8  # Round up to nearest byte
-            return self._value.to_bytes(num_bytes, "big")
-
-        def to_int(self) -> int:
-            """
-            Returns the integer representation of the BitDict.
-            This method provides a way to access the underlying integer value
-            that represents the BitDict's data.
-            Returns:
-                int: The integer value of the BitDict.
-            """
-
-            return self._value
-
-        def to_json(self) -> dict[str, Any]:
-            """
-            Converts the BitDict to a JSON-serializable dictionary.
-            Iterates through the BitDict in reverse order, creating a dictionary
-            where keys are the names of the bitfields and values are their
-            corresponding values. If a value is a BitDict itself, its `to_json`
-            method is called recursively to convert it to a JSON-serializable
-            dictionary as well.
-            Returns:
-                dict[str, Any]: A dictionary representation of the BitDict suitable
-                for JSON serialization.
-            """
-
-            result: dict[str, Any] = {}
-            for name, _ in list(self)[::-1]:  # Use the iterator
-                result[name] = self[name]
-                if hasattr(result[name], "to_json"):
-                    result[name] = result[
-                        name
-                    ].to_json()  # Recurse for nested bitdicts.
-            return result
-
-        def valid(self) -> bool:
-            """Checks if all properties have valid values."""
-            for prop_name, prop_config in self._config.items():
-                if prop_config["type"] == "bitdict":
-                    selector_value = self[prop_config["selector"]]
-                    assert isinstance(
-                        selector_value, int
-                    ), "Selector must be an integer"
-                    if not _is_valid_value(
-                        selector_value, self._config[prop_config["selector"]]
-                    ):
-                        return False
-                    sub_bitdict = self._get_subbitdict(prop_name, selector_value)
-                    if not sub_bitdict.valid():
-                        return False
-                else:
-
-                    value = self[prop_name]
-                    assert isinstance(
-                        value, (int, bool)
-                    ), "Value must be an integer or boolean"
-                    if not _is_valid_value(value, prop_config):
-                        return False
-            return True
-
-        def verify(self) -> bool:
-            """Verifies the BitDict against a verification function.
-            The verification function is a user-defined function that takes a BitDict
-            instance as input and returns a boolean value indicating whether the
-            BitDict is valid. This method calls the verification function with the
-            current BitDict instance and returns the result.
-            Returns:
-                bool: True if the BitDict is valid according to the verification function,
-                False otherwise.
-            """
-
-            return self.__class__.verification_function(self)
-
-    # end class BitDict
-
-    # Set the name of the dynamically created class.
-    BitDict.__name__ = name
-    return BitDict
+__all__ = ["BitDictABC", "bitdict_factory"]
